Fix: Support the case where the credential scheme cannot be deduced
